Login backend and ui setup
